Rename truffle-contract-sources to @truffle/contract-sources
