+++ packages/compile-solidity/new/index.js
@@ -2,7 +2,7 @@ const debug = require("debug")("compile:new"); // eslint-disable-line no-unused-
 const path = require("path");
 const { promisify } = require("util");
 const expect = require("@truffle/expect");
-const findContracts = require("truffle-contract-sources");
+const findContracts = require("@truffle/contract-sources");
 const Config = require("truffle-config");
 const Profiler = require("../profiler");
 const CompilerSupplier = require("../compilerSupplier");
