+++ packages/compile-solidity/package.json
@@ -17,7 +17,7 @@
     "semver": "^5.6.0",
     "solc": "^0.5.0",
     "truffle-config": "^1.1.20",
-    "truffle-contract-sources": "^0.1.5"
+    "@truffle/contract-sources": "^0.1.5"
   },
   "devDependencies": {
     "babel-core": "^6.26.0",
