+++ packages/contract-sources/README.md
@@ -1,2 +1,2 @@
-# truffle-contract-sources
+# @truffle/contract-sources
 Utility for finding all contracts within a directory
