+++ packages/contract-sources/package.json
@@ -1,22 +1,22 @@
 {
-  "name": "truffle-contract-sources",
+  "name": "@truffle/contract-sources",
   "version": "0.1.5",
   "description": "Utility for finding all contracts within a directory",
   "main": "index.js",
   "scripts": {},
-  "repository": "https://github.com/trufflesuite/truffle/tree/master/packages/truffle-contract-sources",
+  "repository": "https://github.com/trufflesuite/truffle/tree/master/packages/contract-sources",
   "keywords": [
     "ethereum",
     "contracts",
     "find",
     "filesystem"
   ],
-  "author": "Tim Coulter <tim.coulter@consensys.net>",
+  "author": "Tim Coulter <tim@trufflesuite.com>",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/trufflesuite/truffle-contract-sources/issues"
+    "url": "https://github.com/trufflesuite/truffle/issues"
   },
-  "homepage": "https://github.com/trufflesuite/truffle-contract-sources#readme",
+  "homepage": "https://github.com/trufflesuite/truffle/tree/master/packages/contract-sources#readme",
   "dependencies": {
     "debug": "^4.1.0",
     "glob": "^7.1.2"
