+++ packages/truffle-core/lib/testing/soliditytest.js
@@ -1,7 +1,7 @@
 const TestCase = require("mocha/lib/test.js");
 const Suite = require("mocha/lib/suite.js");
 const Deployer = require("truffle-deployer");
-const find_contracts = require("truffle-contract-sources");
+const find_contracts = require("@truffle/contract-sources");
 const compile = require("@truffle/compile-solidity/legacy");
 const abi = require("web3-eth-abi");
 const series = require("async").series;
